Renamed urlsanitize module
