+++ corehq/motech/tests/test_requests.py
@@ -13,8 +13,8 @@ from corehq.motech.const import OAUTH2_PWD, REQUEST_TIMEOUT
 from corehq.motech.models import ConnectionSettings
 from corehq.motech.requests import get_basic_requests
 from corehq.motech.views import ConnectionSettingsListView
-from corehq.util.urlsanitize.urlsanitize import PossibleSSRFAttempt
-from corehq.util.urlsanitize.ip_resolver import CannotResolveHost
+from corehq.util.urlvalidate.urlvalidate import PossibleSSRFAttempt
+from corehq.util.urlvalidate.ip_resolver import CannotResolveHost
 from corehq.util.view_utils import absolute_reverse
 
 BASE_URL = 'http://www.example.com/2.3.4/'
