+++ corehq/motech/views.py
@@ -23,7 +23,7 @@ from corehq.motech.forms import ConnectionSettingsForm, UnrecognizedHost
 from corehq.motech.models import ConnectionSettings, RequestLog
 from no_exceptions.exceptions import Http400
 
-from corehq.util.urlsanitize.urlsanitize import PossibleSSRFAttempt
+from corehq.util.urlvalidate.urlvalidate import PossibleSSRFAttempt
 
 
 class Http409(Http400):
